+++ pwsh/CHANGELOG.md
@@ -1,4 +1,11 @@
 # Changelog
 
+## 0.0.2 - 2024-09-08
+ - Improve startup logging.
+ - Improve error feedback.
+ - Add user configurable path per script feature.
+   - Still needs to be under `/share`
+ - Update README.
+
 ## 0.0.1 - 2024-09-07
  - Initial release - should be considered alpha.